backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# Optional but recommended
gunicorn==21.2.0  # For production deployment
pytest==7.4.3  # For testing
black==23.12.0  # Code formatting
flake8==6.1.0  # Linting